Pull Request #54 Corrections: - Extracted TabBar and NavBar creation into a separate structure - TabBarControllerBuilder
